pyproject.toml
[project]
name = "mcp-snowflake-reader"
version = "0.2.0"
description = "Read-only MCP server for Snowflake databases"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Miseon Kim" }
]
dependencies = [
    "mcp>=1.6.0",
    "sqlparse>=0.5.0",
    "python-dotenv>=1.0.0",
    "snowflake-connector-python>=3.10.0",
]

[project.scripts]
mcp-snowflake-reader = "mcp_snowflake_reader.main:main"

[tool.setuptools]
packages = ["mcp_snowflake_reader"]
package-dir = {"" = "src"}
